Update Posts and Comments types and design
The 'comments' type in the posts feature was changed from a simple Comment object to an array of comments. Subsequently, the Post UI was updated to better accommodate the change. Also, a 'User' field was added to the Comment type in the comments feature to include the username of the commenter.