Add post editing functionality
The code changes involve adding the ability to edit posts in the application. An "Edit Post" modal form has been added to the PostList component, which uses the updatePost function from the api. Also, the 'content' property in the updatePost api function has been changed to 'body' to be consistent with the rest of the codebase.