Add dotenv and update dependencies
We introduced dotenv to manage environment variables and updated several dependencies. The "dotenv" package now allows us to define the base URL for API calls in an environment variable. The dependencies "@babel/plugin-proposal-private-property-in-object" and "dotenv" have been updated to the latest versions. A new example file for environment variables (.env.example) is also added and the actual .env file is added to .gitignore for security.